Improve roadmap duration display and tool requirement readability
- Added formatDurationDays to convert roadmap step durations from days to human-readable format (hr/min)
- Replaced raw duration display in RoadmapSection with formatted output
- Fixed incorrect formatting of tool requirement progress (e.g., 0.12 hrs → 7 min)
- Introduced formatHoursMinutes utility for displaying progress as 'hr min / hr min'
- Refactored completed progress display to use consistent formatting

+++ Models/Progression.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct ToolProgress: Codable {
     let toolName: String
-    let requiredHours: Int
+    let requiredHours: Double
     var progress: TimeInterval = 0  // in seconds
 }
